3225 - testable interface for writing files
For example usage of file operations, see filesystem.mu.

Is it ugly that we don't actually write to disk unless we wait for the
writing routine to exit? Maybe there's a nice way to wrap it. At any
rate, all buffering is explicit, which seems a win compared to *nix.