3268 - starting to support separate compilation
Right now Mu has zero dependency knowledge. If anything changes in our
project the C++ compiler has to redo the entire project. This is
unnecessarily slow, and also causes gcc to run out of RAM on puny
machines.

New vision: carve the tangled mu.cc into multiple files.

  includes.h
  types.h
  globals.cc
  globals.h
  one .cc file for each function definition

(This is of course in addition to the already auto-generated test_list
and function_list.)

With this approach changes to functions will only require recompiling
the functions that changed. We'd need to be smart to not rewrite files
that don't change (modulo #line directives).

Any changes to includes/types/globals would still require rebuilding the
entire project. That's the (now greatly reduced) price we will continue
to pay for outsourcing dependency management to the computer.

Plan arrived at after conversation with Stephen Malina.