3287 - drop hacky implicit meaning of CXXFLAGS
Now that we have a new build system we shouldn't need to run unoptimized
just to save time. (Though that's not strictly true; if a change
modifies .build/mu_0.cc which is twice as large as later compilation
units, dropping -O3 shaves 10s off the time for an incremental build.)

Since we don't need to run unoptimized anymore, let's just explicitly
ask for --test-only-app when we need it.