3288 - cleaner heuristic for cleaving
For the last couple of days I've been implicitly thinking in terms of
how many compilation units I want to generate. Might as well make that
explicit and drop the hacky ideas for approximating it.

I tried more timing experiments like the ones in commit 3281.
Conclusion: I can't have the best of both worlds:

  1. Full compilation doesn't take too much longer than with a single
  compilation unit.

  2. Incremental compilation is fast enough that there's negligible
  benefit from dropping optimization.

We're still taking on a 10s hit in full build time.

I care more about not degrading the full compilation too much, since
that gets magnified so much on the Couch's puny server. So we'll just
have to continue using CXXFLAGS=-g when we care to save a few seconds in
incremental compilation time.

A final mystery: the build time increases by 10s with the new heuristic
even though the number of calls to the compiler (and therefore the fixed
cost) is the same. Seems like separating certain functions into
different units is causing the compiler issues. Dropping from 4 to 3
compilation units eliminated the issue.

--- Appendix: Measurements
before:
  full build 4 + test: 42s
  incremental compilation with -O3: varied from 30s for mu_0.cc to 5s for mu_3.cc
    longer times benefitted from dropping -O3

after:
  full build 1 + test: 39s
  full build 2 + test: 41s
  full build 3 + test: 43s
  full build 4 + test: 52s
  full build 5 + test: 53s
  full build 6 + test: 51s
  full build 10 (9) + test: 54s
  full build 20 (16) + test: 58s