3490
Redo commit 3457.

Basically there were 3 unicode characters we changed back then:
  solid horizontal line: 9473 -> 9472
  fuzzy horizontal line: 9480 -> 9548
  fuzzy vertical line: 9482 -> 9550

The solid horizontal line has no issues, so we just redo it here.
For the other two, we'll perform the substitution only when rendering
html. That gives us the best of both worlds: the scenario screens render
right in html, and alt-tabbing continues to be snappy when running the
edit/ app.