3534
Streamline the build process.

It's safest to always:
  a) check if each output is `older_than` the inputs necessary to
  construct it, and
  b) update the output only if something changed.

However:

  i) We don't yet have helpers to do b) in all situations, and
  ii) combining a) and b) can cause `older_than` to spuriously report
  files being updated.

So we'll always run exactly one of a) and b) and err on the side of
keeping the output reliable, at the risk of occasionally updating a file
unnecessarily and triggering unnecessary work downstream. Cross that
bridge when we run into it.