3572 - new way to write server tests
This time we're opening real sockets so we might run into issues on
machines with firewalls. Macs for example flash up a dialog warning
people about a port being opened, though it shuts down immediately if
the test passes.

On the flip side, the test has greater verisimilitude. You don't really
need fakes except when you want to pin down the environment a test runs
in. The only way this test might be flaky is on a machine that has lots
of sockets open (so the random port opened by the test is in use for
something else). That and the firewall concern above. Hmm.