3852
Bugfix of commit 3850 for the sandbox/ app. I'd hoped to just quickly move
past this ugly approach, but a cleaner way is more involved than I thought.

This way is ugly partly because I'm introducing a bunch of conditionals
without testing them. One or more of my additions may well be hiding bugs.
Or I may need to add them in a few other places.

The clean way is to update the fake screen model to accurately mimic the
new real screen, where out of bounds prints aren't silently ignored, and
where scrolling is a fact of life.