3866
Yet another bugfix, this time in just the sandbox/ app:
  open sandbox/ with empty lesson/ directory
  type 'a'
  press backspace
  cursor was not moving left

Now fixed.

Turns out the sandbox/ app hadn't been working right since commit 3854.
(Which ironically was a revert but clearly didn't revert enough; the last
truly good commit was 3823, and we're still clawing our way back to the
sunlight.)

The issue in this case was that commit 3853 disabled update-cursor in some
situations when it shouldn't have. To be safe, just always update-cursor
one very event. I should probably reorganize this in edit/ as well, but
it's not necessary for this particular bug.

---

Incidentally, as part of my git bisecting I realized that the bug fixed
in the trace browser as part of commit 3862 was very old:
  press '/'
  press some key
  press ctrl-u to erase
  press some key
  = out of bounds string access