3954
As a blanket rule, down-arrow now stops scrolling once the bottom margin
comes on screen.

Now that we have page-wise scrolling with ctrl-f/b and line-wise
scrolling with ctrl-s/x, we don't need to conflate scroll positioning
with the arrow keys. And as a result, early students no longer have to
struggle with accidentally scrolling part of the sandbox off the screen
when there's tons of empty space available.

`move-to-next-line` is still super messy and will need further
rethinking, but this commit simplifies the codebase as a whole by
eliminating a couple of historical accidents:

  a) We only introduced scrolling past the bottom of the screen to allow
  more sandboxes to come into view before we had scrolling for the
  sandbox side.

  b) We undid scrolling past the bottom in just the recipe side to allow
  errors to come into view.

Since these historical details are now irrelevant, we no longer need
separate logic for the recipe and sandbox sides, and we don't need to
keep track of the recipe-bottom separate from the bottom margin of
arbitrary editors.