4227 - second example implementing exceptions
I think I like this better. It doesn't violate the type system. Still two
sources of klunkiness:

a) We need to add dead code because we don't consider exceptions when we
add implicit 'return' statements to functions. Should be easy to fix.
b) Still no way to create a generic 'try' function. This will be hard to
fix.