4269 - start validating alloc-ids on lookup
Seems incredible that this is all it took. Needs more testing.

I also need to rethink how we organize our layers about addresses.
Alloc-id stuff is scattered everywhere. The space for alloc-ids is
perhaps unavoidably scattered. Just assume the layout from the start.
But it seems bad that the scenario testing the lookup-time validation is
in the 'abandon' layer when the code is in the 'lookup' layer.