4276 - switching gears to subx
New plan: spend some time learning to program in machine code atop subx,
relying solely on a tiny subset of kernel-provided syscalls. Gradually
introduce helpers.

Helpers we're sure we don't need, so far:
  a) Nested expressions
  b) Garbage collection
  c) One-size-fits-all memory allocation primitive
  d) Function overloading and generics

Helpers we're sure we need, so far:
  a) Dependency-injected versions of syscalls
  b) Tangling directives
  c) Statically checked types

Workflow for a C translator from ascii to binary:
  a) run generated machine code atop subx (unit tests probably go here)
  b) emit machine code packaged as an ELF file
  c) check that the ELF binary runs natively
  d) check that the ELF binary can be unwrapped and run atop subx

This is different from a conventional compiler because the 'HLL' is
unconstrained.
It is also different from Forth given the emphasis on types.

We want a simple stack that also encourages code sharing between
programmers. Conventional languages grow monotonically complex. Forth
discourages code sharing; it is non-trivial to figure out the 'shape' of
data a strange function expects on the stack.