4356 - subx: first program with a data segment
We read() a character from stdin and write() it out to stdout, saving it
to a global variable in between.

ELF binaries are inefficient; you can ask for a low alignment, but the
kernel may not be able to handle it. If you set up a high alignment then
you end up wasting an increasing amount of space in each segment
because of the constraint that the offset bear some relationship with
the loaded address.