4414 - subx: syntax checking
This is a large patch, and there's a few things wrong with it:

a) Helpers are incredibly messy. I want to use has_metadata in layer 24,
but can't since it also does error checking. There must be a better
basis set of primitives for managing metadata.

b) Layer 22 introduces operands for checking, but programs with operands
don't actually run until layer 24. So I can't write non-error scenarios
in layer 22. That seems ugly. But if I try to introduce layer 24 first
there's nothing left to check after it.

I *could* play tricks with ordering layers vs transforms. Mu does that a
bit, but it becomes hard to mess with, so I'm trying to avoid that. My
current plan is for layers within an "abstraction level" to be run in
order. Higher layers will necessarily need to come before lower ones.
But hopefully this level of hierarchy will help manage the chaos.

c) The check for whether an instruction is all hex bytes makes me
nervous. I do want to check that an instruction that's just:
  cd
tells the programmer that an operand is missing. The check I currently
have is likely not perfectly correct.

I *could* put layer 25 in its own commit. But I guess I'm not doing
that now.

We have a new example program: hello world!