4487
Draft attempt at cleaning up warnings, but this isn't quite right.

We still emit warnings for every level-1 scenario, and hiding for each
of them seems painful.

Even if we do that, level-2 scenarios would want to hide level-3 and
over warnings, but *not* level-1 warnings. So we need a cardinal number
rather than booleans.