4567 - support automated tests in SubX
All it takes is to code-generate a simple function called 'run_tests' that
calls all functions starting with 'test_' one by one.

I've temporarily switched the factorial app to run as a test. But that's
temporary, because all the code to print '.' vs 'F' needs to get extracted
out into a helper.