4506
check_ints_equal now prints a newline after the failure message on failure.

We still don't know how to print a final newline after all the tests have
run, for the common case when all tests pass.

Ideally I could just emit a few instructions to `run_tests`. But I'd also
need to add a variable for the newline to the data segment. Or I need some
literal syntax for newlines in strings. We don't have support for backslash-escapes
yet.