4616 - fix subx/examples/ex7
It was broken since I added support for global variables, back on Sep 1.

One other subtle thing I've improved is the name `looks_like_hex_int`.
We can now distinguish in the pack-operands transform between ignoring
'foo' because it doesn't look like a number, and immediately flagging '0xfoo'
as an error because it *should* be a number.