5142
Hoist address computation out of the loop.

I'm giving in to the temptation to optimize here, and violating my own
rule of minimizing local variables by introducing 'curr'. My fig leaf is
that the number of instructions inside the loop goes down, and duplicating
inside the loop may be distracting to readers.