minor style cleanup
I like to be able to jump between functions and tests using '}' in Vim
:)

We also shouldn't put the wrong argument types in the tabular columns.
Just make it obvious they're not in use when instructions don't have a
ModR/M byte.

End-of-line comments are really only useful for instructions with a ModR/M
byte.