.
When rerunning most recently run test, also load the same files as
before. Try the following commands in sequence:

  $ ./run_one_test.sh 056trace.subx test-compute-addresses        # 1
  0 error: missing reference to global 'test-compute-addresses'
  $ ./run_one_test.sh apps/survey.subx test-compute-addresses     # 2
  $ ./run_one_test.sh 056trace.subx trace  # run previous test    # 3

Now command 3 will work, and it will behave identically to command 2.

This way I can switch to another file in Vim and hit `\t` in a random
place in the file and still have the previously run test be rerun.